+++ openapi-cli/src/test/java/io/ballerina/generators/client/BallerinaClientGeneratorTests.java
@@ -16,17 +16,14 @@
  * under the License.
  */
 
-package io.ballerina.generators;
+package io.ballerina.generators.client;
 
 import io.ballerina.compiler.syntax.tree.AnnotationNode;
 import io.ballerina.compiler.syntax.tree.FunctionBodyNode;
-import io.ballerina.compiler.syntax.tree.FunctionSignatureNode;
 import io.ballerina.compiler.syntax.tree.NodeList;
-import io.ballerina.compiler.syntax.tree.ParameterNode;
-import io.ballerina.compiler.syntax.tree.RequiredParameterNode;
-import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.SeparatedNodeList;
 import io.ballerina.compiler.syntax.tree.SyntaxTree;
+import io.ballerina.generators.BallerinaClientGenerator;
+import io.ballerina.generators.OpenApiException;
 import io.ballerina.openapi.cmd.Filter;
 import io.ballerina.openapi.exception.BallerinaOpenApiException;
 import io.ballerina.tools.diagnostics.Diagnostic;
@@ -52,11 +49,10 @@ import java.util.Set;
 import static io.ballerina.generators.BallerinaClientGenerator.extractDisplayAnnotation;
 import static io.ballerina.generators.BallerinaClientGenerator.generatePathWithPathParameter;
 import static io.ballerina.generators.BallerinaClientGenerator.getFunctionBodyNode;
-import static io.ballerina.generators.BallerinaClientGenerator.getFunctionSignatureNode;
 import static io.ballerina.generators.BallerinaClientGenerator.getReturnType;
-import static io.ballerina.generators.TestUtils.compareGeneratedSyntaxTreeWithExpectedSyntaxTree;
-import static io.ballerina.generators.TestUtils.getDiagnostics;
-import static io.ballerina.generators.TestUtils.getOpenAPI;
+import static io.ballerina.generators.common.TestUtils.compareGeneratedSyntaxTreeWithExpectedSyntaxTree;
+import static io.ballerina.generators.common.TestUtils.getDiagnostics;
+import static io.ballerina.generators.common.TestUtils.getOpenAPI;
 
 /**
  * All the tests related to the {@link io.ballerina.generators.BallerinaClientGenerator} util.
@@ -70,14 +66,6 @@ public class BallerinaClientGeneratorTests {
     List<String> list2 = new ArrayList<>();
     Filter filter = new Filter(list1, list2);
 
-//
-//    {"petstore_server_with_base_path.yaml"},
-//    {"petstore_without_operation_id.yaml"},
-//    {"petstore_get.yaml"},
-//    {"openapi_display_annotation.yaml"},
-//    {"header_parameter.yaml"},
-//    {"petstore_post.yaml"},
-//    {"petstore_with_oneOf_response.yaml"}
     @Test(description = "Generate Client for path parameter has parameter name as key word", enabled = true)
     public void generateClientForJira() throws IOException, BallerinaOpenApiException, FormatterException,
             OpenApiException {
@@ -109,30 +97,6 @@ public class BallerinaClientGeneratorTests {
         Assert.assertEquals(generatePathWithPathParameter("/v1/{version}/v2/{limit}"),
                  "/v1/${'version}/v2/${'limit}");
         Assert.assertEquals(generatePathWithPathParameter("/v1/{age}/v2/{name}"), "/v1/${age}/v2/${name}");
-    }
-
-    @Test(description = "Test for generate function signature for given operations")
-    public void getFunctionSignatureNodeTests() throws IOException, BallerinaOpenApiException {
-        OpenAPI openAPI = getOpenAPI(RES_DIR.resolve("swagger/valid_operation.yaml"));
-        FunctionSignatureNode signature = getFunctionSignatureNode(openAPI.getPaths()
-                .get("/products/{country}").getGet());
-        SeparatedNodeList<ParameterNode> parameters = signature.parameters();
-        Assert.assertFalse(parameters.isEmpty());
-        RequiredParameterNode param01 = (RequiredParameterNode) parameters.get(0);
-        RequiredParameterNode param02 = (RequiredParameterNode) parameters.get(1);
-        RequiredParameterNode param03 = (RequiredParameterNode) parameters.get(2);
-
-        Assert.assertEquals(param01.paramName().orElseThrow().text(), "latitude");
-        Assert.assertEquals(param01.typeName().toString(), "float");
-
-        Assert.assertEquals(param02.paramName().orElseThrow().text(), "longitude");
-        Assert.assertEquals(param02.typeName().toString(), "float");
-
-        Assert.assertEquals(param03.paramName().orElseThrow().text(), "country");
-        Assert.assertEquals(param03.typeName().toString(), "string");
-
-        ReturnTypeDescriptorNode returnTypeNode = signature.returnTypeDesc().orElseThrow();
-        Assert.assertEquals(returnTypeNode.type().toString(), "ProductArr|error");
     }
 
     @Test(description = "Tests for returnType")
@@ -173,31 +137,6 @@ public class BallerinaClientGeneratorTests {
                 "response=check self.clientEp-> get(path, accHeaders, targetType = http:Response );returnresponse;}");
     }
 
-    @Test(description = "Tests functionBodyNodes including statements according to the different scenarios",
-            dataProvider = "dataProviderForFunctionBody")
-    public void getFunctionBodyNodes(String yamlFile, String path, String content) throws IOException,
-            BallerinaOpenApiException {
-        Path definitionPath = RES_DIR.resolve(yamlFile);
-        OpenAPI display = getOpenAPI(definitionPath);
-        Set<Map.Entry<PathItem.HttpMethod, Operation>> operation =
-                display.getPaths().get(path).readOperationsMap().entrySet();
-        Iterator<Map.Entry<PathItem.HttpMethod, Operation>> iterator = operation.iterator();
-        FunctionBodyNode bodyNode = getFunctionBodyNode(path, iterator.next());
-        content = content.trim().replaceAll("\n", "").replaceAll("\\s+", "");
-        String bodyNodeContent = bodyNode.toString().trim().replaceAll("\n", "").replaceAll("\\s+", "");
-        Assert.assertEquals(bodyNodeContent, content);
-    }
-
-    @Test(description = "Test openAPI definition to ballerina client source code generation with diagnostic issue",
-            dataProvider = "singleFileProviderForDiagnosticCheck")
-    public void checkDiagnosticIssues(String yamlFile) throws IOException, BallerinaOpenApiException,
-            FormatterException, OpenApiException {
-        Path definitionPath = RES_DIR.resolve("swagger/" + yamlFile);
-        syntaxTree = BallerinaClientGenerator.generateSyntaxTree(definitionPath, filter);
-        List<Diagnostic> diagnostics = getDiagnostics(definitionPath, syntaxTree);
-        Assert.assertTrue(diagnostics.isEmpty());
-    }
-
     @Test(description = "Test openAPI definition to ballerina client source code generation",
             dataProvider = "fileProviderForFilesComparison")
     public void  openApiToBallerinaCodeGenTestForClient(String yamlFile, String expectedFile) throws IOException,
@@ -217,35 +156,6 @@ public class BallerinaClientGeneratorTests {
                 {"uber_openapi.yaml", "uber_openapi.bal"},
                 {"multiple_pathparam.yaml", "multiple_pathparam.bal"},
                 {"covid19_openapi.yaml", "covid19_openapi.bal"}
-        };
-    }
-
-    @DataProvider(name = "singleFileProviderForDiagnosticCheck")
-    public Object[][] singleFileProviderForDiagnosticCheck() {
-        return new Object[][] {
-                {"petstore_server_with_base_path.yaml"},
-                {"petstore_without_operation_id.yaml"},
-                {"petstore_get.yaml"},
-                {"openapi_display_annotation.yaml"},
-                {"header_parameter.yaml"},
-                {"petstore_post.yaml"},
-                {"petstore_with_oneOf_response.yaml"}
-        };
-    }
-
-    @DataProvider(name = "dataProviderForFunctionBody")
-    public Object[][] dataProviderForFunctionBody() {
-        return new Object[][]{
-                {"swagger/header_parameter.yaml", "/pets", "{string path=string`/pets`;" +
-                        "map<string|string[]>accHeaders=" +
-                        "{'X\\-Request\\-ID:'X\\-Request\\-ID,'X\\-Request\\-Client:'X\\-Request\\-Client};" +
-                        "http:Response response=check self.clientEp-> get(path, accHeaders, targetType = " +
-                        "http:Response);returnresponse;}"},
-                {"file_provider/swagger/uber_openapi.yaml", "/history", "{string  path = string `/history`;\n" +
-                        "        map<anydata> queryParam = {offset: offset, 'limit: 'limit};\n" +
-                        "        path = path + getPathForQueryParam(queryParam);\n" +
-                        "        Activities response = check self.clientEp-> get(path, targetType = Activities);\n" +
-                        "        return response;}"},
         };
     }
