+++ openapi-cli/src/test/resources/generators/client/anyOf/ballerina_project/Ballerina.toml
@@ -1,0 +1,5 @@
+[package]
+org= "ballerina"
+name= "test"
+version= "0.0.0"
+
