+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/header_parameter.yaml
