+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/openapi_display_annotation.yaml
