+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/petstore_get.yaml
