+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/petstore_server_with_base_path.yaml
