+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/petstore_server_without_base_path.yaml
