+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/petstore_with_oneOf_response.yaml
