+++ openapi-cli/src/test/resources/generators/client/diagnostic_files/petstore_without_operation_id.yaml
