+++ openapi-cli/src/test/resources/generators/schema/ballerina/openapi_weather_api_schema.bal
