+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema01.bal
