+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema02.bal
