+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema03.bal
