+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema04.bal
