+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema05.bal
