+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema06.bal
