+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema07.bal
