+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema08.bal
