+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema09.bal
