+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema10.bal
