+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema11.bal
