+++ openapi-cli/src/test/resources/generators/schema/ballerina/schema14.bal
