+++ openapi-cli/src/test/resources/generators/schema/swagger/openapi_weather_api.yaml
