+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario01.yaml
