+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario02.yaml
