+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario03.yaml
