+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario04.yaml
