+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario05.yaml
