+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario06.yaml
