+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario07.yaml
