+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario08.yaml
