+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario09.yaml
