+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario10.yaml
