+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario11.yaml
