+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario12.yaml
