+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario13.yaml
