+++ openapi-cli/src/test/resources/generators/schema/swagger/scenario14.yaml
